+++ nix/modules/project-modules.nix
@@ -17,6 +17,8 @@ in
               imported in other flakes.
             '';
             defaultText = lib.literalMD ''
+              TODO: Update the docs
+
               Package and dependency information for this project exposed for reuse
               in another flake, when using this project as a Haskell dependency.
 
@@ -34,26 +36,10 @@ in
 
         config.haskellFlakeProjectModules =
           let
-            defaults = rec {
+            defaults = {
               packages = { pkgs, lib, ... }: withSystem pkgs.system ({ config, ... }: {
                 inherit (config.haskellProjects.default)
                   packages;
-              });
-              # The 'output' module provides both local package and dependency
-              # overrides.
-              output = {
-                imports = [ input local ];
-              };
-              # The 'local' module provides only local package overrides.
-              local = { pkgs, lib, ... }: withSystem pkgs.system ({ config, ... }: {
-                source-overrides =
-                  lib.mapAttrs (_: v: v.root)
-                    config.haskellProjects.default.packages;
-              });
-              # The 'input' module contains only dependency overrides.
-              input = { pkgs, ... }: withSystem pkgs.system ({ config, ... }: {
-                inherit (config.haskellProjects.default)
-                  source-overrides overrides;
               });
             };
           in