+++ nix/modules/project.nix
@@ -6,8 +6,6 @@ let
     types;
   inherit (types)
     raw;
-
-  haskellOverlayType = import ../types/haskell-overlay-type.nix { inherit lib; };
 in
 {
   imports = [
@@ -62,36 +60,6 @@ in
       example = "pkgs.haskell.packages.ghc924";
       default = pkgs.haskellPackages;
       defaultText = lib.literalExpression "pkgs.haskellPackages";
-    };
-    source-overrides = mkOption {
-      type = types.attrsOf (types.oneOf [ types.path types.str ]);
-      description = ''
-        Source overrides for Haskell packages
-
-        You can either assign a path to the source, or Hackage
-        version string.
-      '';
-      default = { };
-    };
-    overrides = mkOption {
-      type = haskellOverlayType;
-      description = ''
-        Cabal package overrides for this Haskell project
-                
-        For handy functions, see 
-        <https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/haskell-modules/lib/compose.nix>
-
-        **WARNING**: When using `imports`, multiple overlays
-        will be merged using `lib.composeManyExtensions`.
-        However the order the overlays are applied can be
-        arbitrary (albeit deterministic, based on module system
-        implementation).  Thus, the use of `overrides` via
-        `imports` is not officiallly supported. If you'd like
-        to see proper support, add your thumbs up to
-        <https://github.com/NixOS/nixpkgs/issues/215486>.
-      '';
-      default = self: super: { };
-      defaultText = lib.literalExpression "self: super: { }";
     };
     autoWire =
       let